Add column-limit with basic strategies.
The column limit penalty, if any, is added to each split cost.

A split can attach a strategy to itself to limit/enchance future splits,
for example if we want to do a one-line block we don't want to allow
splits inside the block. The current implementation has limitations
shown in basic.test.