Upgrade to scalameta 0.22.
This is a combination of 17 commits that were squashed for an easier
rebase:
* (wip) migrates to 0.1.0-RC3+
* RC4: Keyword => KwKeyword
* RC4: position => pos
* RC4: privateTag => /dev/null
* RC3: better Ident extractors
* RC3: isInstanceOf[T <: Token] => is[T]
* RC4: isInstanceOf[T <: Tree] => is[T]
* bugfix
* another bugfix
* RC4: better @classifier
* Fix broken tests.
* reformat

Handle corner case with superfluous parentheses in Term.If.

WIP Upgrade to RC7, something is hanging?

Fix dialects issue, 30 failing tests.

Handle case when splits is empty and empty string.

All tests pass, but performance is ~20x slower:
[info] Benchmark           Mode  Cnt      Score     Error  Units
[info] GenJsCode.scalafmt  avgt   10  30698.572 ± 736.364  ms/op

Upgrade to scala.meta 0.22