Add nailgun interrop to CLI.
I was too excited to release 0.4.5 that I forgot to make sure basic CLI
features like --stdin and --files worked as expected.

This commit reorganizes the CLI so that all files have absolute paths
and all relative paths are read from the injected working directory.

While the nailgun integration is experimental, I want to keep the script
in scalafmt_ng.