Make nested path exclusions working on Windows (#648)
Now paths like target/foo are accepted also on Windows.
The existing unit test has been extended to verify such cases.
By the way this fixed the `scripted` task when running on Windows.

In fact it was possible to use \\ e.g. in paths passed via --exclude,
but it isn't very helpful and convenient in case of projects built on
various platforms.