Issue #607 Add param to don't insert new lines after annotations
There's added noNewlineAfterAnnotation param to OptIn, `stat` test
and the additional check in CliOptionsTest.

The logic related to annotations is located in TreeOps. I changed
it to - when needed - use the alternative approach where we check
whether the line of the beginning of the annotation is the same as
the line of the end of previous one.