Resolve references in config file
In HOCON one might define values using substitutions[1] which are references to
other parts of the configuration tree.

Those references need to be resolved before trying to query the configuration.

Usually substitutions could be resolved by looking at the system environment
variables, but that is explicitly disabled to begin with.

[1]: https://github.com/typesafehub/config/blob/master/HOCON.md#substitutions