Clarify config values & examples some
Present the examples of the configurations in a more explicit, clarified
and format-consistent (oh the irony!) way.

As an example the comments "// ScalaDoc" and "// JavaDoc" weren't clear
if they were just giving an examples of ScalaDoc vs JavaDoc style, or if
they were the valid values to pass to "docstrings".