Fix #798. (#803)
infixSplit is used for both infix applications, where we want to handle
infix operators, and also for assignment, where we don't want to handle
right associative operators.

This commit is a poor fix/hack/workaround for this issue. I would love
to come with a better design/implementation to avoid this duct taping,
but I'm afraid I don't have so much time right now and I want to publish
a fix quickly.