Move all the files!
This commit reorganizes the top-level directories to follow a consistent
naming convention: scalafmt-MODULE. I'm doing this in this PR because
the files have already moved into `shared/` directories so URLs to
previous files will get lost anyways. I like the consistent naming
scheme because it feels easier to get an overview of the project.