Revert "Indent body of vals when unindentTopLevelOperators=true."
This reverts commit b4123dcbf08c56e02b0ebbe7c109d7db1d6b60e8.

This commit introduced a regression for cases like
```scala
// before
val x = a + List(
  1
)
// after
val x = a + List(
    1
  )
```

I believe we either need to properly define what indentation is desired
for unindentTopLevelOperators=true when combined with assignments. I
can't come up with a reasonable definition so I'm personally more
tempted to remove the option altogether. The option is not documented on
the website.