Avoid newlines after yield keyword.
```scala
// Before
for (a <- as)
yield
  Future {
    // ...
  }
// After
for (a <- as)
yield Future {
  // ...
}
```

It's possible to get the old behavior with `newlines.avoidAfterYield = false`.

While working on this change, I observed that
`style.indentYieldKeyword=true` (default) has undesirable (and
unexpected) behavior for val definitions.
```scala
// Before
val x = for (a <- as)
  yield a
// After
val x =
  for (a <- as)
    yield a
```
The new behavior is more consistent with how we handle other val
definition bodies. By forcing a line break before `for`, we make
it easier to associate the `for` keyword with its accompanying `yield`.