Single-arg apply of assign: preserve no-split
This change tries to maintain the existence or absence of split before
the assignment in a single-arg call.

In some cases, this was the way code was formatted already, even for the
config-style configurations.

However, this is also required as a prerequisite for the next commit, to
avoid a couple of test failures where a split was induced by the lambda
change.

The tests added earlier clearly demonstrate the effects of the change.