Router: possibly keep LF before select, and indent
Currently, the code doesn't introduce newlines into select chains even
if their length exceeds the maximum line length.

It also results in missing indent when a comment is in the middle of
such a chain.

Fix both problems, by allowing a newline if one is present in the input.

Fixes #1334.