Router: try break before a comment in apply
We are also moving the optimal break point to before a close parenthesis
if a comment precedes or follows it and doesn't fit.

Since we don't break just before a comment, it occasionally led to very
long lines. While this change can't guarantee the lines won't be longer
than maxColumn either, they will be shorter because we will induce a
break before the close parenthesis:

instead of

  a(b) // comment

we'll format as

  a(
    b
  ) // comment