Router: don't break comment after a trailing comma
The problem with trailing commas is possible idempotency due to their
being added at the end of the process. Formatting decisions are made in
the absence of that comma, which has two aspects:

* some rules (like the one being addressed here) treat tokens after a
  comma differently than if the comma hadn't been there;
* when adding it leads to a too-long line, the next pass over the code
  will probably result in a different formatting.

Ultimately, the right approach would be to create a rewrite rule which
would be enabled based on the 'trailingCommas' setting, and which would
append or remove those commas before formatting process starts.