Router bugfix: newline split for assign in apply
For cases when leaving the assign value on the same line as the opening
parenthesis would lead to exceeding maxColumn, let's add the option to
break line. The other newline split below, one that was disabled for the
single-arg assign case, doesn't work as well, hence we need to add this
explicit alternative.

Instead of

a(b =
  c
)

this will allow

a(
  b = c
)

when "a(b" is too long.