Router bugfix: keep single-line comment split
Even if we are bin-packing, we shouldn't fold single-line comments which
were previously detached. First, the general rule is breaks for comments
are normally maintained; second, even for bin-packing, comments would
frequently be used to separate groups within literal lists, and removing
the break is undesirable. This was also not requested in #431, the issue
which was fixed with the introduction of this behaviour.