RedundantBraces bugfix: skip non-infix with infix
That is, if a block is a parameter of an infix expression, only rewrite
it if it contains a further infix expression, or something simple.

This takes care of incorrectly removing braces around functions (either
explicit, or with placeholder parameters), or Unit expressions (such as
Defn), and other complex situations surrounding infix.

Fixes #1631 #1633.