RedundantBraces bugfix: skip infix if joins lines
We must deal with infix expressions differently; for instance,
   ...
 } && ...
with our prior approach would be rewritten as
  ...
  && ...
which is an invalid expression.

On the other hand, if we were to remove the preceding newline to get
  ... && ...
the downside is that we usually don't modify line breaks around infix
expressions.

Additionally, if there's a comment just before that newline, that could
result in another invalid expression:
    ... // comment
  } && ...
cannot become
    ... // comment && ...

So the decision is to remove braces in infix expressions only if the
closing brace is preceded by a non-whitespace token on the same line
so that deleting the newline wouldn't be needed.