Router: allow single-line curly lambdas
The rule applies only if
 style.newlines.alwaysBeforeCurlyBraceLambdaParams = false
 style.newlines.afterCurlyLambda = never, or preserve and no line break

Also, make sure to clarify the other space rule for lambda to break on
arrow before breaking on the closing brace, so that it doesn't end up
with a lone break at the end.

This is a prerequisite to adding the rewrite rule requested in #1027.
Also, fixes #965.

For alwaysBeforeCurlyBraceLambdaParams = false:
  a map { x => x }
instead of
  a map { x =>
    x
  }

For alwaysBeforeCurlyBraceLambdaParams = true:
  a map {
    x => x
  }
instead of
  a map {
    x =>
      x
  }