Trailing commas: add a rewrite rule, remove first
In order to avoid idempotency issues triggered by modified commas at the
final stage of formatting, let's remove them in the beginning to have a
clean format, and append as necessary.

This rule is only applied if the TrailingCommas configuration takes one
of the mutating settings, i.e. always or never.

Fixes #1667.