Split: add a convenience method to ignore split
Similar to the multitude of withXXX methods, let's add a method to apply
further filtering to a split. Also, for those of us who prefer positive
logic to the negative logic of the ignoreIf field, call it onlyIf().

Modify a few invocations to demonstrate usage in cases where it becomes
more compact and readable.