TreeOps bugfix: find Init param group correctly
Unlike Apply (which gets a single set of arguments, with currying parsed
as successive Apply trees), Init gets all its parameter groups.

However, the owner of the token on the right is not always a parameter,
it could also be the entire Init tree, such as when that token is, say,
a comment. So look for tokens, not trees.