FormatOps: find the exact param group for each "("
Rather than returning the concatenation of all parameter groups, return
the specific one.

With that, we can now definitively determine the number of arguments in
the specific parameter group, rather than the total number across all
groups as we did before.

That is, in

  def f(a: Int)(b: String, c: Boolean): Unit

we used to return 3 in getApplyArgs for both '(', but would now return 1
and 2, respectively.