TokenOps bugfix: correctly identify a blank line
While we might be able to ignore some line breaks in Scala, a blank line
(or at least two consecutive newlines) is treated differently. Let's add
a method to FormatToken to identify it, and fix a bug in TokenOps where
input with three consecutive newlines was handled as one, not two.