Router: extract, simplify enumerator splits logic
There's no need for the delayed arrow logic as in case clauses, as the
case rule applies it on "case", while the enumerator arrow was applied
just before arrow instead of simply doing right after.

Also, move the arrow/val generator matches down, just before the comment
matches, so the arrow rules is not superseded by a possible comment.