Align strip margin: allow disabling the alignment
Previously, using `assumeStandardLibraryStripMargin = true` (which means
the `.stripMargin` method was not redefined by the user and thus assigns
no special meaning to indentation preceding the `|` character) resulted
in aligning the pipe character with the end of the opening triple-quote.

Now, one can optionally set `align.stripMargin = false` to use a 2-space
indentation offset relative to the start of the opening line instead.

// With align.stripMargin = true
val x = """
          |Hello world!
          |""".stripMargin
// With align.stripMargin = false
val x = """
  |Hello world!
  |""".stripMargin