Sunset editions
Editions are a failed experiment. My original motivation to add editions
was to make it possible to make improvements to the default formatting
behavior without introducing a new configuration option in every PR.
This made sense when Scalafmt was evolving slowly, but lately the
project has been moving much faster thanks to the heroic efforts of
Albert Meltzer and other contributors.

I'm happy that Scalafmt is still evolving and I like many of the new
changes. However, the latest improvements don't respect the `edition`
setting, which is understandable because the changes are large.

I don't want to remove editions because that would break user's
configuration files. Instead, this commit just updates the website to
say that editions are no longer used.