BinPack: allow single-line formatting of literals
Now that we have fileOverride which can be applied to tests, one thing
that is frequently needed in tests is setting up a number of instances
or parameterized tests with multiple parameters, leading to extremely
vertical formatting for what is essentially not code but configuration.

For that case, allow a special form of binpacking which simply formats
all literal arguments on one line, even if it overflows maxColumn.