ScalafmtSession: add an interface for batch format
If we need to format a number of files using the same exact scalafmt
configuration, let's create a session. This will also allow us to
filter inputs based on the project files which was not available and
lead to dynamic runner loading and formatting all files, including
those not covered by the configuration.