FormatWriter bugfix: comment format on lookaround
Our comment-space-and-asterisk-matching regular expression was modified
to avoid lookaround constructs, but that inadvertently introduced a bug.

When the comment line contains just a single asterisk, the character
that follows it is the newline; that newline should be included as the
last character of one regex match and then as the first character of the
next match. Alas, that doesn't happen.

Therefore, restore the positive lookaround constructs to avoid this bug.