Scaladoc: upgrade to scalameta 4.3.15
The new version of scalameta fixes a few bugs (code blocks with an empty
line in the middle, lists starting at the beginning of a comment, links
followed by punctuation). Modify the formatter to consider these.

However, don't output lists on top line for style=AsteriskSpace; in this
style, the top line has a different margin than subsequent ones, while
lists are very sensitive to margin space. (The same would also apply to
style=Asterisk, but in that style, the top line is always empty.)