ScalafmtDynamic: throw if can't resolve config
Intellij invokes the createSession method directly and doesn't expect
the null value we return there, so throw ConfigError instead.

Similarly, in the Runner, change the implementation of MyInstanceSession
to verify the config first, so that the config parsing exception is not
thrown for every file we are trying to format.