Add failing test with comment before first select
Also, modify existing tests to ensure that comments in select chains are
not provided without indentation, as this has been historically handled
differently to be consistent with IDE behaviour commenting out code.

Finally, add test with docstring after block; since the fix we introduce
in the next commit will shift indent from FormatWriter to Router, that
will expose a feature(?) in FormatWriter which ignored a blank line in
such a case and output only a single newline.