ScalafmtConfig: validate once in the decoder only
There are multiple places where a configuration instance is created, and
some of them are not really appropriate for validation, such as StyleMap
where custom configuration copies are created based on comments in the
code, or to support binPack.literalArgumentLists.

Therefore, let's just validate the initial configuration in the decoder
and not the custom adjustments.