FormatOps: allow excluding end in exclude ranges
For SingleLineBlock, the left token is checked in the provided exclude
range sets; in some cases, say when the end token refers to the closing
brace, we might want to make sure the closing brace is not followed by
a newline. For that, the end token needs to be excluded from the range.