Add support to expand command-line arguments with @ symbol
Fixes #2145. Previously, it was not possible to invoke the Scalafmt
command-line tool with a large number of files as arguments because it
would exceed the operating system limit for argument lengths. Now,
clients can pass in arbitrarily long arguments by writing them
first to a file.
```
$ cat /path/to/arguments
/path/to/File.scala
$ scalafmt @/path/to/arguments
```
The command above is equivalent to running "scalafmt /path/to/file.scala".