Scala 3: require custom dialect to parse Scala 3 syntax.
Previously, the default dialect supported a subset of Scala 3 syntax.
This caused problems because the Scala 3 parsing support in Scalameta
still has bugs that may introduce regressions for Scala 2 users (example
https://github.com/scalameta/scalameta/issues/2191). This commit
changes the default dialect to only recognize Scala 2 syntax. Scala 3
users need to add `runner.dialect = Scala3` to their configuration file
in order to recognize Scala 3 syntax.

The long-term goal is to understand Scala 3 syntax out of the box. This
change is only meant as a temporary situation while the Scala 3 support
in Scalameta stabilizes.