Scala 3: add support for extension methods (#2298)
* Scala 3: add support for extension methods

Previously, Scalafmt didn't support formatting extension methods.
This commit fixes that issue for code using braces. There's still no
support for indentation syntax (which is still being designed and is
likely to change soon).

* Change extension to soft keyword

`extension` is currently a soft keyword and we can reuse the SoftKeywors class from scalameta to check this. I also added a number of tests for extension being used as a non-keyword and also for extensions with comments before def

Co-authored-by: Tomasz Godzik <tomek.godzik@gmail.com>