Newlines: add forceBeforeMultilineAssign
Deprecate the short-lived newlines.beforeMultilineDef, since it is now
covered by this option, together with newlines.beforeMultiline.

Also, copy template member tests with this new parameter set to true.
Since the functionality hasn't been implemented yet, the setting causes
no difference in formatting, and the results are identical for now.

Finally, add a few test cases with infix expression as the body, and a
few with various levels of nesting.