Router: use extractors on meta to get extra values
Using extractors on the entire format token, especially when it includes
a match on either left or right token, is inefficient. However, in most
cases an extractor match on the meta field is equivalent and allows us
to extract the value once instead of twice (once in match and a second
time in the case body).