TreeOps: don't extract single stmt in block
This method is used to identify the points where new statements start,
specifically in order to figure out where one statement ends and a new
one begins.

With the exception of special handling for lambdas, let's not "start" a
new statement when it's the only one, as there's no previous statement
to separate from.

This allows with rewriting of optional braces, when the LeftBrace token
is dropped but the block around it remains, which leads to triggering
an earlier, incorrect rule.