FormatTokens: allow to look up an ealier token
In preparation to the change which potentially removes tokens from the
sequences, we might end up looking up a removed token. For that, we will
need to be able to choose the token after or the token before.

Let's return the preceding token by default and add an explicit method
to get the subsequent one instead.