FormatTokensRewrite: base to rewrite rules on FT
In some cases, when the rewrites are not too complex, that is, mostly
replace or remove some tokens, we can avoid reparsing the resulting code
by modifying the resulting FormatTokens (while keeping the underlying
tress unchanged).

Let's implement the framework to apply such rules, and later we'll add
the rules themselves.