FormatOps: mimic {} pairs with optional braces
With optional braces, we are missing the physical indication that the
formatter has come to expect with some of its optimizations.

The BestFirstSearch algorithm has an optimization which recurses into a
block and doesn't backtrack. Some split policies skip matching pairs of
braces when applying a single-line constraint.

In the absence of any other indicators (like Indent/Outdent tokens), we
need to identify cases where braces are optional.