Use sbt-native-image plugin to build native image.
Previously, the build used sbt-assembly to generate the native-image
binary. This made it necessary to start sbt with a pre-installed GraalVM
version. With sbt-native-image, the correct GraalVM version is installed
automatically.

Also, this commit tweaks the GraalVM native-image linking options to
force the `AbsoluteFile` class to initialize a runtime. This should fix
a bug when passing in relative paths to native-image binaries.