Add Scala Native support
Project structure was changed in compliance with sbt-crossproject.

googlecode's diffutils had to be removed, as Scala Native does not
support java libraries. In its place, part of diffutils was
reimplemented in scala as part of scalafmt-cli.

For similar reasons, the java interfaces were reimplemented in scala for
Scala Native.

Since Scala Native does not yet support glob expressions, a translation
function to regex was added, heavily based on a linked stack-overflow
answer. Similarly, Scala's Process is replaced with java stdlib's
ProcessBuilder, as Process uses concurrency.

SN's regex implementation is currently based on the re2 instead of the
standard java regexes. This means there are a number of differences
between them, most notably the lack of backtracking in Scala Native.
To accomodate that, some regex function reimplementations were done to
accomodate the use of the existing scalafmt regexes in Scala Native.

Some operations concerning tests were slightly modified to work on
Native. For example, renaming a file to an existing directory path could
throw errors. This is fine, as it does not concern the core of the test,
only things like cleanup.

Assumptions about dynamic tests not working on native were added.
Scala Native does not support URLs and only has basic reflection support
comparable to the one of GraalVM, so the used and expected version of
scalafmt must match here as well.

Term Display was made single-threaded for the Scala Native
implementation, as it does not support concurrency. Since the new
implementation caused display artifacts on JVM, the old multithreaded
one was kept there.

Scala Native setup and testing was added to the CI. The windows was
temporarily disabled as it does not pass all of the tests.

scala-native sbt command was also added for easier scala-native cli
building.

Co-authored-by: Tomasz Godzik <tgodzik@users.noreply.github.com>