FormatOps vertical multiline: improve relay policy
Ensure that `allOwners` not only exclude the current paren owner but all
the previous ones, too, making sure they are correctly sorted by their
position.

Then reverse their order for Policy.Relay, thus ensuring that the left,
or `before`, part of every relay is a simple policy, with the complex
Relay in the `after` position.