TreeOps: fix handling implicit group keyword
Earlier, we would only handle the first `implicit` differently if none
of the params have an explicit `implicit` keyword.

In reality, we should relax the check and do this if not all params have
an explicit keyword. The idea being, if they all do, then the initial
one belongs to the first parameter rather than the group.