RemoveScala3OptionalBraces: handle infix on rbrace
Instead of handling a specific owner of the left brace (previously, it
was a Term.Match only), let's instead look for infix when we encounter
the right brace instead, and ensure that the next token is not the op
of an infix.

This also handles the case when Term.Match was not an LHS but in the
middle of an infix chain.