FormatWriter: accumulate align shift correctly
Alignment logic accumulates alignment shift for matching lines, with the
assumption that the extra space will definitely be added.

However, we don't add extra spaces to no-split cases, instead attempting
to delay them until the next token... but if the next token is also a
no-split, then the extra alignment space is lost, leading to incorrect
formatting.

One such example is `(` which typically has no spaces before or after.