FormatOps: outindent after fewer braces in infix
When a fewer braces call is used before an infix operator, that operator
may not be indented as it will otherwise be interpreted as part of the
fewer-braces argument.

This fix leaves a few cases (notably, many for newlines.source=unfold)
still broken, but that requires a change to the parser which will be
released in the new version of scalameta.