FormatWriter: fix end marker handling
The current issue is that the span is computed using `leftLineId` of the
token just before and the token at the end. However, `leftLineId` refers
to the beginning of the token, thus multiline comments are incorrectly
accounted.

Instead, use the token at the beginning of the span and increase by 1.