BestFirstSearch: switch to Long as State key
State column is unlikely to exceed 24 bits (or 16+ mln bytes), but even
in cases when it might, a given token (that is, at a given State depth)
cannot conceivably be formatted at two different column values which are
a multiple of 2^24 bytes apart.

Hence, let's form state key by keeping indentation (just like before)
at 8 bits, limiting column to 24 bits and combining with depth in the
remaining 32 bits (and this combining is what ensures no hash conflict).