Router: no space around `{` to be changed to `(`
Let's not use the Space split when we know that it will be removed along
with the brace being change to paren.

This will allow us to avoid non-idempotent formatting problems as these
spaces might cause the code to overflow the line narrowly, which would
not happen without the spaces.