CliOptions: write info to console instead of err
Currently, info is written to out if out is not used for result output,
and to err otherwise.

Let's enhance this approach as follows: if out is not used, then we'll
write into to out, as before; otherwise, before resorting to err, check
if the console is available and use it then.