Modification: add SpaceOrNoSplit, handle State etc
This new modification would serve as a space unless we encounter, on the
same line, a terminating token, in which case it will become a NoSplit.

This will allow us to re-implement conversion of braces to parens for
one-line apply without having to worry about possibly non-idempotent
additional formatting rules.