improvement: Pin image versions
Without the change it seems it's not possible to use scalafmt on older ubuntu images due to older gcc version. This should not be an issue normally as users can just update. For windows it's just a precaution since that is also the latest image and windows-2019 seems way too old.