BestFirstSearch: relax cost check w/ optimal token
Rather than requiring that state cost remain unchanged with the latest
split, let's limit it to checking that the original cost of the split
(before any penalties may have been applied) is zero, and that the
state does not overflow.

This leads to extremely rare formatting changes (none on scala, spark,
scalafmt, scalameta repos) and on average over a 5% improvement in the
BestFirstSearch algorithm efficiency.

This approach protects optimal-token intentions from penalties which
are designed to rank different states, and is a follow-on to #4851.