Add working "get started" example
At the moment, the "Get started" button on the homepage links to the
Installation documentation, but doesn't include instructions on
setting Scalafmt up for a project.

It's tough to discover that there are two required pieces of
configuration, and what the values of these should be. Even the
configuration page (which addresses more complex examples) there's no
clear of example of a minimal working config definition that teams can
start from. This means Scalafmt is only accessable to engineers that
already know how it works, or that have somewhere else to copy a
working config file from.

This change adds a note to the top of the Installation documentation,
because:

1. this page is linked from "Get started" so it is the first thing people see
2. It's reasonable to interpret "Installing" to mean setting up Scalafmt on the project

We use the `@STABLE_VERSION@` reference so that these docs will always
point to the correct version of Scalafmt. This is especially important
because there are two different versions of Scalafmt (`scalafmt-core`
/ `sbt-scalafmt`), which is a detail that someone just setting up IDE
formatting won't need to understand.