Add working example to the configuration docs
Changes the first example in the configuration example so that it
includes the mandatory config options, to better serve as a starting
point for customisation.

We also link to the Configuration section at the top of the
Installation page, since the Installation documentation serves as a
"Get started" introduction.

Taken together, this should help new Scala engineers get started with
Scalafmt.