Splits: handle braces within if/for/while better
Previously, we'd assume that a brace within `if` or `while` contains the
entire condition and let the brace rules handle it; however, this brace
could enclose only a part of the condition and hence needs to be handled
differently then.

Additionally, we ignored the danglingParentheses.ctrlSite parameter in
these cases as well.