FormatWriter: again fix infix align owner
Keep the owner unchanged but apply this special logic mapping operator
(Term.Name) to parent infix tree (Term.ApplyInfix) only when checking
if the operator token is to be aligned using configured matchers.

If an infix expression starts with a multiline subexpression but ends
with a single line of tokens to be aligned, then subsequent single-line
statements are properly aligned with it.

However, in the opposite case, when an infix expression starts with a
single line of tokens to be aligned and continues with a multiline part,
then it was not aligned with previous statements.

By keeping the owner, we fix this inconsistency.

Includes undo of #4904.